Fix Database

+++ dotnet_project/Controllers/CartController.cs
@@ -28,7 +28,7 @@ namespace dotnet_project.Controllers
             return View("~/Views/Checkout/Index.cshtml");
         }
 
-        public async Task<IActionResult> Add(int Id)
+        public async Task<IActionResult> Add(long Id)
         {
             ProductModel product = await _dataContext.Products.FindAsync(Id);
             List<CartItemModel> cart = HttpContext.Session.GetJson<List<CartItemModel>>("Cart") ?? new List<CartItemModel>();
